+++ src/tests/tests.js
@@ -1,0 +1,99 @@
+import test from 'tape';
+import {uglify, dashed, dotted, normalize, longDistance, extensionNumber, groupTwo, groupFour} from '../index';
+
+test('Return a uglified phone number', t => {
+	let result = uglify('555-444-1111');
+
+	t.equal(result, '5554441111', `Returned uglify format: ${result}`);
+
+	result = uglify('555444-1111');
+	t.equal(result, '5554441111', `Returned uglify format: ${result}`);
+	t.end();
+});
+
+test('Return a groupedTwo format phone number', t => {
+	let result = groupTwo('44332211');
+
+	t.equal(result, '44 33 22 11', `grouped numbers into 4 groups of two: ${result}`);
+	result = groupTwo('5544332211');
+
+	t.equal(result, '55 44 33 22 11', `grouped 10 digit number to groups of two: ${result}`);
+	t.end();
+});
+
+test('Return a groupedFour format phone number', t => {
+	let result = groupFour('44332211');
+
+	t.equal(result, '4433 2211', `grouped numbers into 2 groups of four: ${result}`);
+	result = groupFour('5544332211');
+
+	t.equal(result, '55 4433 2211', `grouped 10 digit number to group of 2 4 4: ${result}`);
+	t.end();
+});
+
+test('Return a dashed format phone number', t => {
+	let result = dashed('5554441111');
+
+	t.equal(result, '555-444-1111', `Returned dashed format: ${result}`);
+	result = dashed('555.444.1111');
+
+	t.equal(result, '555-444-1111', 'Converted dotted formatting to dashed');
+	result = dashed('4441111');
+
+	t.equal(result, '444-1111', `Returned short dashed format: ${result}`);
+	result = dashed('444.1111');
+
+	t.equal(result, '444-1111', `Returned short dashed format: ${result}`);
+	t.end();
+});
+
+test('Return a dotted format phone number', t => {
+	let result = dotted('5554441111');
+
+	t.equal(result, '555.444.1111', `Returned dotted format: ${result}`);
+	result = dotted('555-444-1111');
+
+	t.equal(result, '555.444.1111', 'Converted dashed to dotted format');
+	result = dotted('4441111');
+
+	t.equal(result, '444.1111', `Returned dotted format: ${result}`);
+	result = dotted('444-1111');
+
+	t.equal(result, '444.1111', `Returned dotted format: ${result}`);
+	t.end();
+});
+
+test('Return a normal format phone number', t => {
+	const result = normalize('5554441111');
+
+	t.equal(result, '(555) 444-1111', `Returned normalize format: ${result}`);
+	t.end();
+});
+
+test('Return a long distance format', t => {
+	let result = longDistance('15554441111', 'dotted');
+
+	t.equal(result, '1+555.444.1111', `Returned longdistance with dotted format: ${result}`);
+	result = longDistance('205554441111', 'dotted');
+
+	t.equal(result, '20+555.444.1111', `Returned longdistance with dotted format: ${result}`);
+	result = longDistance('2695554441111', 'dotted');
+
+	t.equal(result, '269+555.444.1111', `Returned longdistance with dotted format: ${result}`);
+	result = longDistance('17875554441111', 'dotted');
+
+	t.equal(result, '1-787+555.444.1111', `Long Distance formatted with long country code: ${result}`);
+	result = longDistance('4414815554441111', 'dotted');
+
+	t.equal(result, '44-1481+555.444.1111', `Long Distance formatted with extra long country code: ${result}`);
+	t.end();
+});
+
+test('Return an extension format', t => {
+	let result = extensionNumber('55544411118989', 'dashed');
+
+	t.equal(result, '555-444-1111 x 8989', `Returned extension with dashed format: ${result}`);
+	result = extensionNumber('5554441111899', 'dashed');
+	t.equal(result, '555-444-1111 x 899', `Returned extension with dashed format: ${result}`);
+	t.end();
+});