+++ changelog.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## v1.4.0
+
+> - `longDistance` now supports country codes of all shapes and sizes
+>   - This includes: `X`, `XX`, `XXX`, `X-XXX`, `XX-XXXX`
+> - Code optimizations
+> - Some re organization of the strcuture
+> - Added two new formats
+>   - These new formats are **NOT** supported for use with `longDistance` or `extension`
+>   - `groupTwo` which gives out a format of `XX XX XX XX` supports both 8 and 10 character numbers
+>   - `groupFour` Which gives out a format of `XXXX XXXX` supports both 8 and 10 character numbers
+>     - 10 character numbers output `XX XXXX XXXX`
+
 ## v1.3.0
 > - Code optimizations
 > - Better UMD support